# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: T.A.R.S. Evaluation Engine API
  description: |
    Real RL agent evaluation with regression detection and baseline management.

    The Evaluation Engine provides:
    - Multi-environment agent evaluation
    - Performance regression detection
    - Nash equilibrium scoring for multi-agent conflicts
    - Baseline management with versioned history
    - Integration with AutoML pipeline and HyperSync

    ## Authentication
    Most endpoints require JWT Bearer authentication. Admin endpoints require
    the `admin` role. Viewer endpoints require at least `viewer` role.

    ## Rate Limiting
    - Public endpoints: 30 requests/minute
    - Authenticated endpoints: 10 requests/minute (auth operations)
    - Evaluation endpoints: 5 requests/minute

  version: v1.0.0-rc2
  contact:
    name: T.A.R.S. Development Team
    url: https://github.com/veleron-dev-studios/tars
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8099
    description: Local development
  - url: https://eval-engine.tars.svc.cluster.local:8099
    description: Kubernetes cluster (internal)
  - url: https://api.tars.example.com/eval-engine
    description: Production (via Ingress)
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - auth
      summary: Authenticate user and get JWT tokens
      description: Exchange username/password for access and refresh tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: developer
                password:
                  type: string
                  format: password
                  example: dev_password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token obtained from /auth/login.

        The token must include:
        - sub: username
        - role: viewer|developer|admin
        - exp: expiration timestamp

        Example: `Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`
  schemas:
...
    TokenResponse:
      type: object
      required:
        - access_token
        - refresh_token
        - token_type
      properties:
        access_token:
          type: string
          description: JWT access token (60 min expiry)
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refresh_token:
          type: string
          description: JWT refresh token (7 day expiry)
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        token_type:
          type: string
          example: bearer
        expires_in:
          type: integer
          description: Access token expiry in seconds
          example: 3600
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error type
          example: validation_error
        message:
          type: string
          description: Human-readable error message
          example: Invalid agent_type value
        details:
          type: object
          description: Additional error context
          additionalProperties: true
        trace_id:
          type: string
          description: Distributed trace ID for debugging
          example: "abc123def456"

  # ==========================================
  # Common Responses
  # ==========================================
  responses:
...
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: unauthorized
            message: "Invalid or expired token"
...
    TooManyRequests:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: rate_limit_exceeded
            message: "Rate limit exceeded: 10 requests/minute"
            details:
              retry_after_seconds: 45
...